refactor(zk-explorer): Split tabs into separate components
